+++ src/libAtomVM/opcodesswitch.h
@@ -63,18 +63,18 @@ extern "C" {
 #ifdef IMPL_EXECUTE_LOOP
 
 #if AVM_NO_JIT
-#define SET_ERROR(error_type_atom)           \
-    ctx->exception_class = ERROR_ATOM;       \
-    ctx->exception_reason = error_type_atom; \
+#define SET_ERROR(error_type_atom)                \
+    context_set_exception_class(ctx, ERROR_ATOM); \
+    ctx->exception_reason = error_type_atom;      \
     ctx->exception_stacktrace = stacktrace_create_raw(ctx, mod, pc - code, ERROR_ATOM);
 #elif AVM_NO_EMU
-#define SET_ERROR(error_type_atom)           \
-    ctx->exception_class = ERROR_ATOM;       \
-    ctx->exception_reason = error_type_atom; \
+#define SET_ERROR(error_type_atom)                \
+    context_set_exception_class(ctx, ERROR_ATOM); \
+    ctx->exception_reason = error_type_atom;      \
     ctx->exception_stacktrace = stacktrace_create_raw(ctx, mod, (const uint8_t *) native_pc - (const uint8_t *) mod->native_code, ERROR_ATOM);
 #else
 #define SET_ERROR(error_type_atom)                                                          \
-    ctx->exception_class = ERROR_ATOM;                                                      \
+    context_set_exception_class(ctx, ERROR_ATOM);                                           \
     ctx->exception_reason = error_type_atom;                                                \
     if (mod->native_code) {                                                                 \
         ctx->exception_stacktrace = stacktrace_create_raw(ctx, mod,                         \
@@ -85,18 +85,18 @@ extern "C" {
 #endif
 
 #if AVM_NO_JIT
-#define SET_ERROR_MFA(error_type_atom, m, f, a) \
-    ctx->exception_class = ERROR_ATOM;          \
-    ctx->exception_reason = error_type_atom;    \
+#define SET_ERROR_MFA(error_type_atom, m, f, a)                 \
+    context_set_exception_class_use_live_flag(ctx, ERROR_ATOM); \
+    ctx->exception_reason = error_type_atom;                    \
     ctx->exception_stacktrace = stacktrace_create_raw_mfa(ctx, mod, pc - code, ERROR_ATOM, m, f, a);
 #elif AVM_NO_EMU
-#define SET_ERROR_MFA(error_type_atom, m, f, a) \
-    ctx->exception_class = ERROR_ATOM;          \
-    ctx->exception_reason = error_type_atom;    \
+#define SET_ERROR_MFA(error_type_atom, m, f, a)                 \
+    context_set_exception_class_use_live_flag(ctx, ERROR_ATOM); \
+    ctx->exception_reason = error_type_atom;                    \
     ctx->exception_stacktrace = stacktrace_create_raw_mfa(ctx, mod, (const uint8_t *) native_pc - (const uint8_t *) mod->native_code, ERROR_ATOM, m, f, a);
 #else
 #define SET_ERROR_MFA(error_type_atom, m, f, a)                                                          \
-    ctx->exception_class = ERROR_ATOM;                                                                   \
+    context_set_exception_class_use_live_flag(ctx, ERROR_ATOM);                                          \
     ctx->exception_reason = error_type_atom;                                                             \
     if (mod->native_code) {                                                                              \
         ctx->exception_stacktrace = stacktrace_create_raw_mfa(ctx, mod,                                  \
@@ -115,8 +115,8 @@ extern "C" {
     goto handle_error;
 
 #define RAISE_ERROR_MFA(error_type_atom, m, f, a) \
-    SET_ERROR_MFA(error_type_atom, m, f, a)     \
-        goto handle_error;
+    SET_ERROR_MFA(error_type_atom, m, f, a)       \
+    goto handle_error;
 
 #define VM_ABORT() \
     goto do_abort;
@@ -1268,11 +1268,11 @@ static void destroy_extended_registers(Context *ctx, unsigned int live)
         }                                                                                       \
     }
 
-#define RAISE_ERROR_FROM_HELPER(error_type_atom)    \
-    do {                                            \
-        ctx->exception_class = ERROR_ATOM;          \
-        ctx->exception_reason = (error_type_atom);  \
-        return term_invalid_term();                 \
+#define RAISE_ERROR_FROM_HELPER(error_type_atom)           \
+    do {                                                   \
+        context_set_exception_class(ctx, ERROR_ATOM);      \
+        ctx->exception_reason = (error_type_atom);         \
+        return term_invalid_term();                        \
     } while (0)
 
 #ifndef AVM_NO_EMU
@@ -1294,6 +1294,19 @@ static void destroy_extended_registers(Context *ctx, unsigned int live)
         } else {                                                          \
             SCHEDULE_WAIT(mod, pc);                                       \
         }                                                                 \
+    }
+
+#define PROCESS_MAYBE_TRAP_RETURN_VALUE_RESTORE_PC_INDEX_ARITY(return_value, rest_pc, m, i, a)          \
+    if (term_is_invalid_term(return_value)) {                                                           \
+        if (UNLIKELY(!context_get_flags(ctx, Trap))) {                                                  \
+            term module_atom;                                                                           \
+            term function_atom;                                                                         \
+            module_get_imported_function_module_and_name_atoms((m), (i), &module_atom, &function_atom); \
+            pc = rest_pc;                                                                               \
+            HANDLE_ERROR_MFA(module_atom, function_atom, (a));                                          \
+        } else {                                                                                        \
+            SCHEDULE_WAIT(mod, pc);                                                                     \
+        }                                                                                               \
     }
 
 #define PROCESS_MAYBE_TRAP_RETURN_VALUE_LAST(return_value)      \
@@ -1350,7 +1363,11 @@ static void destroy_extended_registers(Context *ctx, unsigned int live)
 #endif
 
 #define HANDLE_ERROR()                                                  \
-    ctx->exception_stacktrace = stacktrace_create_raw(ctx, mod, pc - code, ctx->exception_class);  \
+    ctx->exception_stacktrace = stacktrace_create_raw(ctx, mod, pc - code, context_exception_class(ctx));  \
+    goto handle_error;
+
+#define HANDLE_ERROR_MFA(m, f, a)                                                                                            \
+    ctx->exception_stacktrace = stacktrace_create_raw_mfa(ctx, mod, pc - code, context_exception_class(ctx), (m), (f), (a)); \
     goto handle_error;
 
 #define VERIFY_IS_INTEGER(t, opcode_name, label)           \
@@ -2259,7 +2276,7 @@ loop:
                         case NIFFunctionType: {
                             const struct Nif *nif = EXPORTED_FUNCTION_TO_NIF(func);
                             term return_value = nif->nif_ptr(ctx, arity, x_regs);
-                            PROCESS_MAYBE_TRAP_RETURN_VALUE_RESTORE_PC(return_value, orig_pc);
+                            PROCESS_MAYBE_TRAP_RETURN_VALUE_RESTORE_PC_INDEX_ARITY(return_value, orig_pc, mod, index, arity);
                             x_regs[0] = return_value;
                             if (ctx->heap.root->next) {
                                 if (UNLIKELY(memory_ensure_free_with_roots(ctx, 0, 1, x_regs, MEMORY_FORCE_SHRINK) != MEMORY_GC_OK)) {
@@ -4057,9 +4074,9 @@ wait_timeout_trap_handler:
 
                 #ifdef IMPL_EXECUTE_LOOP
                     TRACE("raise/2 stacktrace=0x%" TERM_X_FMT " exc_value=0x%" TERM_X_FMT "\n", stacktrace, exc_value);
-                    ctx->exception_class = stacktrace_exception_class(stacktrace);
+                    context_set_exception_class(ctx, stacktrace_exception_class(stacktrace));
                     ctx->exception_reason = exc_value;
-                    ctx->exception_stacktrace = stacktrace_create_raw(ctx, mod, saved_pc - code, ctx->exception_class);
+                    ctx->exception_stacktrace = stacktrace_create_raw(ctx, mod, saved_pc - code, context_exception_class(ctx));
                     goto handle_error;
                 #endif
 
@@ -6699,7 +6716,7 @@ wait_timeout_trap_handler:
                                 ex_class != THROW_ATOM)) {
                         x_regs[0] = BADARG_ATOM;
                     } else {
-                        ctx->exception_class = x_regs[0];
+                        context_set_exception_class(ctx, x_regs[0]);
                         ctx->exception_reason = x_regs[1];
                         ctx->exception_stacktrace = x_regs[2];
                         goto handle_error;
@@ -7720,16 +7737,16 @@ bs_match_jump_to_fail:
 #ifdef IMPL_EXECUTE_LOOP
 #ifndef AVM_NO_EMU
 do_abort:
-        ctx->exception_class = ERROR_ATOM;
+        context_set_exception_class(ctx, ERROR_ATOM);
         ctx->exception_reason = VM_ABORT_ATOM;
 #endif
 
 handle_error:
         {
-            x_regs[0] = ctx->exception_class;
+            x_regs[0] = context_exception_class(ctx);
             x_regs[1] = ctx->exception_reason;
             x_regs[2] = ctx->exception_stacktrace;
-            ctx->exception_class = term_nil();
+            context_set_exception_class(ctx, term_nil());
             ctx->exception_reason = term_nil();
             ctx->exception_stacktrace = term_nil();